@@ -3,12 +3,17 @@
   "version": "0.0.5",
   "description": "Text selection library.",
   "main": "lib/selektr.js",
+  "jsnext:main": "es/selektr.js",
   "scripts": {
     "clean": "rimraf lib",
     "lint": "eslint src",
-    "build": "babel src --out-dir lib",
+    "build": "npm run build:commonjs && npm run build:es && npm run build:umd && npm run build:umd:min",
+    "build:commonjs": "BABEL_ENV=commonjs babel src --out-dir lib",
+    "build:es": "BABEL_ENV=es babel src --out-dir es",
+    "build:umd": "mkdir -p dist && BABEL_ENV=commonjs browserify -t babelify src/selektr.js --standalone Selektr | derequire > dist/selektr.js",
+    "build:umd:min": "mkdir -p dist && BABEL_ENV=commonjs browserify -t babelify src/selektr.js --standalone Selektr | derequire | uglifyjs -c -m > dist/selektr.min.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "prepublish": "npm run clean && npm run build"
+    "prepublish": "npm run clean && npm run lint && npm run build"
   },
   "repository": {
     "type": "git",
@@ -26,15 +31,39 @@
   },
   "homepage": "https://github.com/lohfu/selektr#readme",
   "dependencies": {
-    "descendants": "0.0.3",
-    "dollr": "0.0.4",
-    "lodash": "^4.9.0"
+    "dollr": "0.0.5",
+    "lodash": "^4.13.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.7.5",
-    "eslint": "^2.7.0",
-    "eslint-config-airbnb": "^6.2.0",
-    "eslint-plugin-react": "^4.3.0",
-    "rimraf": "^2.5.2"
+    "babel-cli": "^6.9.0",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-check-es2015-constants": "^6.8.0",
+    "babel-plugin-transform-es2015-arrow-functions": "^6.8.0",
+    "babel-plugin-transform-es2015-block-scoped-functions": "^6.8.0",
+    "babel-plugin-transform-es2015-block-scoping": "^6.9.0",
+    "babel-plugin-transform-es2015-classes": "^6.9.0",
+    "babel-plugin-transform-es2015-computed-properties": "^6.8.0",
+    "babel-plugin-transform-es2015-destructuring": "^6.9.0",
+    "babel-plugin-transform-es2015-duplicate-keys": "^6.8.0",
+    "babel-plugin-transform-es2015-for-of": "^6.8.0",
+    "babel-plugin-transform-es2015-function-name": "^6.9.0",
+    "babel-plugin-transform-es2015-literals": "^6.8.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.8.0",
+    "babel-plugin-transform-es2015-object-super": "^6.8.0",
+    "babel-plugin-transform-es2015-parameters": "^6.9.0",
+    "babel-plugin-transform-es2015-shorthand-properties": "^6.8.0",
+    "babel-plugin-transform-es2015-spread": "^6.8.0",
+    "babel-plugin-transform-es2015-sticky-regex": "^6.8.0",
+    "babel-plugin-transform-es2015-template-literals": "^6.8.0",
+    "babel-plugin-transform-es2015-typeof-symbol": "^6.8.0",
+    "babel-plugin-transform-es2015-unicode-regex": "^6.8.0",
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.1",
+    "derequire": "^2.0.3",
+    "eslint": "^2.10.2",
+    "eslint-config-airbnb-base": "^3.0.1",
+    "eslint-plugin-import": "^1.8.0",
+    "rimraf": "^2.5.2",
+    "uglify-js": "^2.6.2"
   }
 }
